Commit final

+++ fonctions.h
@@ -13,7 +13,7 @@ void calcul_accumulation(float *terrain_local, float *dir, float *acc, int nb_ba
 int coords_to_indice(int x,int y, int nBande,int taille_bande,int nb_cols);
 int chercher_min_bord(int x, int y);
 int chercher_min(int x, int y,float *terrain_local, int nBande, int taille_bande, int nb_cols,float nodata);
-void rassembler(float *terrain_local, int *dir, int nb_bandes, int taille_bande, int nb_cols, float no_value);
+void rassembler(float *terrain_local, float *dir, int nb_bandes, int taille_bande, int nb_cols, float no_value);
 
 float* verifier_voisins(int x, int y, float *acc, int *dir, int nBande, int taille_bande, int nb_cols);
 #endif //MNT_DIRECTION_ET_ACCUMULATION_FONCTIONS_H
